feat: Add ButtonTop component
The code changes in this commit add a new component called "ButtonTop" to the application. This component is responsible for rendering a button that appears when the user scrolls down the page. It allows the user to quickly navigate back to the top of the page with a smooth scrolling animation.

This commit aims to improve the user experience by providing an easy way for users to navigate back to the top of the page, especially on long pages with a lot of content.