feat: Add utility function for merging CSS class names
This commit adds a new utility function called "cn" to the application's utils module. The "cn" function takes in multiple CSS class names as arguments and merges them using the "clsx" and "twMerge" libraries. This function will be used to generate the final CSS class name for components in the application.

The addition of this utility function aims to enhance the application by providing a convenient way to merge CSS class names and improve the overall code organization and reusability.